refactor: layout to be server component
